backgroundTasks/nameBgTask/requirements.txt
Django
django-background-tasks

